// src/interfaces/isupplier.ts
export interface GetSupplierDTO {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    address: string;
  }
  
  export interface AddSupplierDTO {
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    address: string;
  }
  